Fixes writing to private channels
When reading or writing from private channels the bot dies
because it only check the list of "channels" inside information and
the it ignores "groups" which is where the private channels' data is
stored.

This PR adds a check on the groups which is readily available inside the
information object when reading and writing a message.