Clarify a few die() errors, fix missing conversations
By default conversations.list (and similar functions) only return 100 entries and a base64 encoded pagination cursor to be used on a future request for the next 100 entries.

Because of the way this module operates it's not an easy change to implement pagination at the moment. Instead I've set the limit parameter to an obnoxiously high number (999999999). Additionally, I've populated the 'types' parameter so that the module will now receive *ALL* conversations rather than dying when dragged into a non-public(im, private_channel, mpim) conversation.