Avoid getting killed by a bug in Slack.
When using shared channels it is possible to receive a user_change event for a
user that is on a different team.

Additionally, there is currently a bug that causes that event to include the
shared channel ID, even if you aren't in it, and even if it's a private channel.

That means conversations.info will return an negative response and we die().

Rather than doing that, warn about it and include the channel ID.

Slack engineers have been notified about this bug but indicated it could be
some time before it is resolved.